% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{biomes}
\alias{biomes}
\title{Biomes of the world}
\usage{
data(biomes)
}
\description{
Shapefile of the world's biomes extracted from Olson's ecoregions of the worlds.
}
\details{
This shapefile contains information about the world's biomes and 
was derived from the terrestrial ecoregions of the world'a shapefile.
}
