% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{eez}
\alias{eez}
\title{Outline of the world's EEZs}
\usage{
data(eez)
}
\description{
Outline of the exclusive economic zone of each country
}
\details{
Current outline of the exclusive economic zone of each country, Version ...
}
