% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{eez_boundaries}
\alias{eez_boundaries}
\title{EEZ boundaries}
\usage{
data(eez_boundaries)
}
\description{
Shapefile of the eez boundaries
}
\details{
This shapefile delineates the maritime boundaries of the world.
}
