% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{package}
\name{geodat package}
\alias{geodat package}
\alias{geodatpackage}
\title{geodat - R package for easy access to various geographical data}
\description{
R package for easy access to various geographical data
}
\author{
RS-eco
}
\keyword{package}
