% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{lme}
\alias{lme}
\title{Large Marine Ecosystems}
\usage{
data(lme)
}
\description{
Shapefile of the world's large marine ecosystems (LME)
}
\details{
This shapefile delineates the world's large marine ecosystems
}
