% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{marinerealms}
\alias{marinerealms}
\title{Marine realms}
\usage{
data(marinerealms)
}
\description{
Shapefile of the world's marine realms
}
\details{
Shapefile of the world's marine realms
}
