% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{maritime_boundaries}
\alias{maritime_boundaries}
\title{Maritime boundaries}
\usage{
data(maritime_boundaries)
}
\description{
Shapefile of the maritime boundaries
}
\details{
This shapefile delineates the maritime boundaries of the world.
}
