% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{meow}
\alias{meow}
\title{Marine ecoregions of the world}
\usage{
data(meow)
}
\description{
Dataset on the marine ecoregions of the world
}
\details{
This shapfile delineates the marine ecoregions of the world.
}
