% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{teow}
\alias{teow}
\title{Terrestrial ecoregions of the world}
\usage{
data(teow)
}
\description{
Shapefile of the world's terrestrial ecoregions
}
\details{
This shapefile contains information about the world's terrestrial ecoregions
}
