% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{zoorealms}
\alias{zoorealms}
\title{Zoogeographic realms of the world}
\usage{
data(zoorealms)
}
\description{
Shapefile of the world's zoogeographic realms
}
\details{
This shapefile contains information about the world's zoogeographic realms.
}
