% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodat-package.R
\docType{data}
\name{zooregions}
\alias{zooregions}
\title{Zoogeographic regions of the world}
\usage{
data(zooregions)
}
\description{
Shapefile of the world's zoogeographic regions
}
\details{
This shapefile contains information about the world's zoogeographic regions.
}
